Enemies, and Player health update
added script for enemies to follow and deal damage to the player upon collision. Worked on the health bar baut the image in the health must be set to square to work I could'nt find a square asset and I think it is  a library error but other than that it should work.